Crate event_udp_to_kafka

event_udp_to_kafka

This module listens to UDP data received on a socket, converts this to flatbuffers-encoded messages, and produces these messages to a Kafka topic (usually _rawEvents).

Modules

config

data_processing

Utilities for converting UDP bytes to flatbuffers-encoded messages.

gps_time

Utilities for GPS timestamps as transmitted over UDP.

metrics

Utilities for exposing Prometheus-compatible metrics

testing

Testing & benchmarking utilities.

udp_message

Utilities for interpreting headers from a UDP message.

Structs

Args

Command-line arguments for the event_udp_to_kafka.

WiringConfigRecord

Wiring table information.

Functions

read_csv

udp_process

Listen to a UDP socket and produce messages onto the output Kafka topic forever.