event_udp_to_kafka::packet_formats

Trait TestablePacketFormat

pub trait TestablePacketFormat: PacketFormat {
    const FAKE_EVENT_TOF: i32;
    const FAKE_EVENT_PIXEL: i32;

    // Required methods
    fn make_fake_event() -> Vec<u8> ;
    fn make_fake_board_specific_header_data() -> Vec<u8> ;
}

Testing utilities for a packet format. These may be used by unit tests and benchmarks, but are not used at runtime.

Required Associated Constants

const FAKE_EVENT_TOF: i32

const FAKE_EVENT_PIXEL: i32

Required Methods

fn make_fake_event() -> Vec<u8>

Manufacture a single fake event. The data contained by this fake event is arbitrary, but should be valid for the type of board. This data is used for benchmarking and testing, so should aim to be representative of a typical real event.

The combination of make_fake_event and make_fake_board_specific_header_data should produce an event with tof FAKE_EVENT_TOF and pixel FAKE_EVENT_PIXEL.

fn make_fake_board_specific_header_data() -> Vec<u8>

Return a board-specific header data suitable for decoding messages from this board.

This should be representative of a real configuration; for example, if the board would typically have multiple channels, this should use a non-zero channel.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".